% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignSum.R
\name{AlignSum}
\alias{AlignSum}
\title{Align GWAS summary stat for the reference population to the target population.}
\usage{
AlignSum(sum_target, sum_ref, SplitSum = TRUE)
}
\arguments{
\item{sum_ref}{The GWAS summary statistics for the reference population.
The most commonly used reference population is European.}

\item{SplitSum}{Execute the SplitSum() function. Default = TRUE.}

\item{sum_tar}{The GWAS summary statistics for the target population.}
}
\value{
data.frame object with ref population GWAS summary statistics
aligned with target GWAS summary statistics. The resulting global variable is
named "sum_com". If SplitSum = TRUE, the SplitSum() and WriteSplitTables()
functions will be executed and the following global variables will be created;
"sum_ref" for the reference snps and "sum_target" for the target snps created
by SplitSum().
}
\description{
Align GWAS summary statistics for the reference population to the target population.
The summary statistics data must have similar formats and include the
following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect
allele. BETA is the regression coefficients for linear regression, log-odds
ratio for logistic regression. SE is the standard error for BETA.
}
\examples{
data.dir <- "data/"
sum_EUR <- fread(paste0(data.dir,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data.dir,"AFR_sumdata.txt"),header=T)
AlignSum(sum_tar = sum_AFR, sum_ref = sum_EUR, SplitSum=TRUE)
}
