% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEBEffectSize.R
\name{CalculateEBEffectSize}
\alias{CalculateEBEffectSize}
\title{CalculateEBEffectSize}
\usage{
CalculateEBEffectSize(bfile,prs_tune,plink_list,memory,threads,
out_prefix,results_dir,params_farm)
}
\arguments{
\item{bfile}{Plink binary files for the target test population.}

\item{snp_ind}{Column name from prs_mat (output from PRSscore()) with best
performance against the tuning data set.}

\item{plink_list}{List of plink data.frame and files from PreparePlinkFiles()}

\item{memory}{Plink "--memory" parameter. Defaults to 8000 mb}

\item{threads}{Plink "--threads" parameter. Defaults to 2}

\item{out_prefix}{Prefix for exported files. Recommended when plink files
are divided by chromosome or chunks, i.e., "chr1" or "chr2", and CTSLEB
script is dispatched in parallel.}

\item{results_dir}{Folder or directory to export files.}

\item{params_farm}{List of plink parameters produced from SetParamsFarm()}

\item{plink2_exec}{Plink2 binary execute file. Default "plink2".}
}
\description{
Performs the entire Empirical-Bayes estimation of effect sizes
workflow as outlined in step2 of the vignette
}
\keyword{EBayes}
