% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEBEffectSizeMulti.R
\name{CalculateEBEffectSizeMulti}
\alias{CalculateEBEffectSizeMulti}
\title{CalculateEBEffectSize}
\usage{
CalculateEBEffectSize(bfile,prs_tune,plink_list,memory,threads,
out_prefix,results_dir,params_farm)
}
\arguments{
\item{ref_names}{The names of the other anescestries for GWAS summary statistics}

\item{sum_ref_list}{A list that contains the GWAS summary statistics for multiple populations.}

\item{bfile}{Plink binary files for the target test cohort.}

\item{snp_list, }{Object created by the RunClump() function.}

\item{plink_list}{List of plink data.frame and files from PreparePlinkFiles()}

\item{results_dir}{Folder or directory to export files.}

\item{out_prefix}{Prefix for exported files. Recommended when plink files
are divided by chromosome or chunks, i.e., "chr1" or "chr2", and CTSLEB
script is dispatched in parallel.}

\item{threads}{Plink "--threads" parameter. Defaults to 2}

\item{memory}{Plink "--memory" parameter. Defaults to 8000 mb}

\item{params_farm}{List of plink parameters produced from SetParamsFarm()}

\item{sum_target}{The GWAS summary statistics for the target population. The data needs to have following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect allele. BETA is the regression coefficients for linear regression, log-odds ratio for logistic regression. SE is the standard error for BETA.}

\item{best_snps_set, }{Object created by the GetSNPSet() function.}

\item{plink2_exec}{Plink2 binary execute file. Default "plink2".}
}
\description{
Performs the entire Empirical-Bayes estimation of effect sizes
workflow as outlined in multi-ancestry section of the vignette
}
\keyword{EBayes}
