% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateQRange.R
\name{CreateQRange}
\alias{CreateQRange}
\title{Create the q range file used for plink2}
\usage{
CreateQRange(x)
}
\arguments{
\item{pthres}{the p value thresholds for selecting SNPs}
}
\value{
the q range file
}
\description{
Create the q range file used for plink2
}
