% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimCT.R
\name{dimCT}
\alias{dimCT}
\title{dimCT}
\usage{
dimCT(plink19_exec, params_farm)
}
\arguments{
\item{plink19_exec}{Plink binary execute file. Default "plink".}

\item{plink2_exec}{description}

\item{results_dir}{description}

\item{sum_target}{description}

\item{sum_ref}{description}

\item{ref_plink}{description}

\item{target_plink}{description}

\item{test_target_plink}{description}

\item{out_prefix}{description}

\item{r2_vec}{description}

\item{wc_base_vec}{description}

\item{memory}{description}

\item{threads}{description}

\item{params_farm}{list of parameters generated by SetClumpFarm().}
}
\description{
This function performs the entire two-dimensional clumping and thresholding
workflow (dimCT) as outlined in step1 of the vignette
}
\examples{
data.dir <- "data/"
temp.dir <- test/temp.dir
system("mkdir -p test/temp.dir")
sum_EUR <- fread(paste0(data.dir,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data.dir,"AFR_sumdata.txt"),header=T)
ref_split_file <- paste0(temp.dir,"sum_ref.txt")
target_split_file <- paste0(temp.dir,"sum_target.txt")
Eur_plinkfile <- paste0(data.dir,"EUR_ref_chr22")
Afr_plinkfile <- paste0(data.dir,"AFR_ref_chr22")
ref_out <- "test/temp/ref",
target_out <- "test/temp/target"
params_farm <- SetParamsFarm(plink19exec = /Apps/plink1.9/plink,
                           wc_base_vec = c(50,100),
                           r2_vec = c(0.01,0.05,0.1,0.2,0.5,0.8),
                           threads = 2,
                           mem = 8000,
)
dimCT(plink19_exec=plink19_exec,
      sum_target = sum_AFR,
      sum_ref = sum_EUR,
      ref_plink= Eur_plinkfile,
      target_plink= Afr_plinkfile,
      ref_split_file = ref_split_file,
      target_split_file = target_split_file,
      ref_clump_out=ref_out,
      target_clump_out=target_out,
      params_farm = params_farm)
}
\keyword{clump}
\keyword{plink1.9}
