% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBayesPostMean.R
\name{EBayesPostMean}
\alias{EBayesPostMean}
\title{EBayesPostMean}
\usage{
EBayesPostMean(clump_info, snp_set)
}
\arguments{
\item{clump_info}{description}

\item{snp_set}{description}
}
\description{
This function performs the entire Empirical-Bayes estimation of effect sizes
workflow as outlined in step2 of the vignette
}
