% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBpost.R
\name{EBpost}
\alias{EBpost}
\title{Estimate Bayesian posterior meanusing the empirical Bayes algorithm}
\usage{
EBpost(unique_infor, SNP_set)
}
\arguments{
\item{unique_infor}{the unique_infor from PreparePlinkFile output, which contains information for all SNPs after clumping step}

\item{SNP_set}{the SNP set from CT step for estimating the covariance matrix for the prior distribution}
}
\value{
Bayesian postier mean for all SNPs after-clumping
}
\description{
Estimate Bayesian posterior meanusing the empirical Bayes algorithm
}
