% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatePrior.R
\name{EstimatePrior}
\alias{EstimatePrior}
\title{Estimate the covariance matrix for the prior distribution used in EB step}
\usage{
EstimatePrior(SNP_set)
}
\arguments{
\item{SNP_set}{the SNP set from CT step for estimating the covariance matrix for the prior distribution}
}
\value{
the covariance matrix estimate
}
\description{
Estimate the covariance matrix for the prior distribution used in EB step
}
