% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatePriorMulti.R
\name{EstimatePriorMulti}
\alias{EstimatePriorMulti}
\title{Estimate the covariance matrix for the prior distribution used in EB step}
\usage{
EstimatePriorMulti(SNP_set, other_ans_names, sum_com)
}
\arguments{
\item{SNP_set}{the SNP set from CT step for estimating the covariance matrix for the prior distribution}
}
\value{
the covariance matrix estimate
}
\description{
Estimate the covariance matrix for the prior distribution used in EB step
}
