% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFinalBetas.R
\name{ExtractFinalBetas}
\alias{ExtractFinalBetas}
\title{Clean the PRS's for the Tune and Validation set for the Super Learner step.}
\usage{
ExtractFinalBetas(
  Tune_PRS,
  Predicted_Tune_Y,
  prs_mat_eb,
  unique_infor_post,
  pthres
)
}
\arguments{
\item{Tune_PRS}{Cleaned Tune PRS's that were used in the Super Learner step. Remove IID and FID, i.e. the first two columns.}

\item{Predicted_Tune_Y}{Raw predicted tune response from the Super Learner.}

\item{prs_mat_eb}{prs_mat_eb from the EB step.}

\item{unique_infor_post}{Global environment object from prior steps.}

\item{pthres}{pthres vector set in earlier steps as a parameter.}
}
\value{
Named list of Cleaned Tune PRS and Cleaned Validation PRS.
}
\description{
Clean the PRS's for the Tune and Validation set for the Super Learner step.
}
