% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_CombinePRS.R
\name{helper_CombinePRS}
\alias{helper_CombinePRS}
\title{Helper function for PRSscore()}
\usage{
helper_CombinePRS(scores, pthres, prs_p_other_)
}
\arguments{
\item{scores}{description}

\item{pthres}{description}

\item{prs_p_other_}{description}
}
\value{
'prs_mat' global variable. Matrix of combined PRSs with clumping and
pthres as column name
}
\description{
Combines PRS
Not exported
}
