% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_CreateQRange.R
\name{helper_CreateQRange}
\alias{helper_CreateQRange}
\title{Helper function for PreparePlinkFile()}
\usage{
helper_CreateQRange(x)
}
\arguments{
\item{x}{pthres}
}
\value{
CreateQRange() object
}
\description{
Calls CreateQRange() with either the default pthres or user input values
Not exported
}
