% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_ebscore_loop.R
\name{helper_ebscore_loop}
\alias{helper_ebscore_loop}
\title{helper_ebscore_loop}
\usage{
helper_ebscore_loop(
  plink2_exec,
  bfile,
  eb_plink_list,
  plink_list,
  pthres,
  threads,
  memory,
  results_dir,
  out_prefix = as.null(),
  params_farm = as.null()
)
}
\arguments{
\item{plink2_exec}{description}

\item{bfile}{description}

\item{eb_plink_list}{description}

\item{plink_list}{description}

\item{pthres}{Default c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)}

\item{threads}{Default 2}

\item{memory}{Default 8000}

\item{results_dir}{description}

\item{out_prefix}{description}

\item{params_farm}{description}
}
\description{
Helper for function for PRSscoreEbayes()
}
\keyword{clump}
\keyword{plink1.9}
