% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pheno.R
\name{helper_pheno}
\alias{helper_pheno}
\title{Helper function for PRS_tune()}
\usage{
helper_pheno(pheno_file, pheno_df)
}
\value{
'prs_mat' global variable. Matrix of combined PRSs with clumping and
pthres as column name
}
\description{
Select file or dataframe source for phenotype y_tun dataframe.
Not exported
}
