% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prefix.R
\name{helper_prefix}
\alias{helper_prefix}
\title{Helper function}
\usage{
helper_prefix(out_prefix, ebayes = FALSE)
}
\arguments{
\item{out_prefix}{description}

\item{ebayes}{Default FALSE}
}
\description{
This function creates the plink out prefix for clumping, scoring and EBayes
scoring. Not exported
}
\keyword{plink}
\keyword{prefix}
