% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_PreparePlinkFile.R
\name{helper_PreparePlinkFile}
\alias{helper_PreparePlinkFile}
\title{Helper function for dimCT()}
\usage{
helper_PreparePlinkFile(plink_list, results_dir)
}
\arguments{
\item{plink_list}{description}

\item{results_dir}{description}
}
\value{
list
}
\description{
Inputs list object from PreparePlinkFile() and creates global variables
and write.tables.
q_range from Prepare
Not exported
}
