% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_qual_order.R
\name{helper_qual_order}
\alias{helper_qual_order}
\title{Helper function for Super_split. Order qual_tune vector}
\usage{
helper_qual_order(x, y)
}
\arguments{
\item{x}{mtx correlation matrix}

\item{y}{qual_tune vector from helper_qual_tune()}
}
\value{
vector
}
\description{
Helper function for Super_split. Order qual_tune vector
}
