% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_qual_tune.R
\name{helper_qual_tune}
\alias{helper_qual_tune}
\title{Helper function for Super_split. Create r2_tune vector}
\usage{
helper_qual_tune(x, x_pheno, pthres, r2_vec, wc_base_vec, pheno_format = 1)
}
\arguments{
\item{x}{vector of tune phenotype values}

\item{pheno_format}{tune phenotype format}
}
\value{
list
}
\description{
Helper function for Super_split. Create r2_tune vector
}
