% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_score_loop.R
\name{helper_score_loop}
\alias{helper_score_loop}
\title{Helper function for PRSscore()}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, score_farm)
}
\arguments{
\item{plink2_exec}{Plink binary execute file. Default "plink2".}

\item{bfile}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{q_range_file}{File name and location of written q_range dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'q_range_file'}

\item{score_col_nums}{number of columns in scores dataframe produced by
PreparePlinkfile().}

\item{scores_file}{File name and location of written scores dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'scores_file'}

\item{p_values}{p_values dataframe produced by PreparePlinkfile()}

\item{pthres}{vector of p_value thresholds used to select snps for PRS.
Default vector name is 'pthres' which is generated by PreparePlinkfile()}

\item{threads}{maximum number of concurrent threads}

\item{memory}{primary workspace memory}

\item{out}{output folder and file prefix for Plink2. Default name prs_p_other_
<pthres column number>.<range name>.sscore}
}
\description{
This function iterates the plink2score function through pthres and p_values
}
\examples{
PRSscore(plink2_exec = "plink2 ",
            bfile = bfile,
            q_range_file = q_range_file,
            score_col_nums = ncols,
            scores_file = scores_file,
            p_values = p_values,
            pthres = pthres,
            threads = 4,
            memory = 8000,
            out,
            scoreFarm = scoreFarm)
}
\keyword{plink2}
\keyword{score}
