% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_SplitSum.R
\name{helper_SplitSum}
\alias{helper_SplitSum}
\title{Helper function for AlignSum()}
\usage{
helper_SplitSum(sum_com, results_dir = results_dir)
}
\value{
SplitSum() and WriteSplitTables() objects and files
}
\description{
SplitSum = TRUE, call the SplitSum() and WriteSplitTables() functions and
create the sum_com global variable object.
SplitSum = FALSE, only create the sum_com global variable object.
Not exported
}
