% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plink19Clump.R
\name{Plink19Clump}
\alias{Plink19Clump}
\title{Plink19Clump}
\usage{
Plink19Clump(plink19_exec, clump, bfile, out, params_farm)

Plink19Clump(plink19_exec,bfile,clump,clump_p1,clump_r2,clump_kb,
threads,memory,out,params_farm)
}
\arguments{
\item{plink19_exec}{Plink binary execute file. Default "plink".}

\item{bfile}{Prefix for plink binary set (prefix.bed, prefix.bim, prefix.fam).}

\item{clump}{PLINK-format association report(s) (text files with a header
line, a column containing variant IDs, and another column
containing p-values).}

\item{clump_p1}{Significance threshold for a SNP to be included as an index SNP.
Default = 1 is used to include all SNPs for clumping.}

\item{clump_r2}{LD r^2 threshold for clumping. SNPs with a higher r^2
with the index SNPs will be removed.}

\item{clump_kb}{Physical distance threshold for clumping. SNPs
within the radius of the index SNP are considered for clumping.
For PRS, clump_kb = base_window_size/clumping_r_square so that lower
clump_r2 thresholds have a larger clumping window size.}

\item{threads}{Maximum number of concurrent threads.}

\item{memory}{Primary workspace memory.}

\item{out}{Output folder and file prefix for Plink1.9. Output files will end
with .clumped}
}
\description{
This function sets the standard options LD clumping with plink1.9.
If additional options are required, please use the source() command
as demonstrated in the vignette.
}
\keyword{clump}
\keyword{plink1.9}
