% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2score.R
\name{plink2score}
\alias{plink2score}
\title{Generate PRS scores with Plink2}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, score_farm)
}
\arguments{
\item{plink2_exec}{Plink binary execute file. Default "plink2".}

\item{bfile}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{score_col_nums}{number of columns in scores dataframe produced by
PreparePlinkfile().}

\item{results_dir}{output folder forvPlink2. Results will appear as results_dir/
temp/prs_p_other_<pthres index>.<range name>.sscore}

\item{pthres_idx}{pthres vector index number}

\item{threads}{maximum number of concurrent threads}

\item{memory}{primary workspace memory}

\item{q_range_file}{File name and location of written q_range dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'q_range_file'}

\item{scores_file}{File name and location of written scores dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'scores_file'}

\item{p_values}{p_values dataframe produced by PreparePlinkfile()}

\item{prs_farm}{prs_farm object name}
}
\value{
plink2 score files
}
\description{
Performs PRS scoring with plink2 for all combinations of clumpings and p-value
threshold combinations
}
\keyword{plink2}
\keyword{score}
