% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlinkFileEB.R
\name{PreparePlinkFileEB}
\alias{PreparePlinkFileEB}
\title{Prepare the files for PLINK2 to calculate PRSs}
\usage{
PreparePlinkFileEB(snp_list, unique_infor_post, post_beta_mat)
}
\arguments{
\item{snp_list}{the snp_list result from the two-dimensional clumping}

\item{sum_com}{the sum_com result from the AlignSum function.}
}
\value{
the score_file, p_value_file and p_other_file for the two-dimensional thresholding
}
\description{
Prepare the files for PLINK2 to calculate PRSs
}
