% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_Clean.R
\name{PRS_Clean}
\alias{PRS_Clean}
\title{Clean the PRS's for the Tune and Validation set for the Super Learner step.}
\usage{
PRS_Clean(Tune_PRS, Tune_Y, Validation_PRS, family = "Gaussian")
}
\arguments{
\item{Tune_PRS}{PRS's for the Tune Data built using the prs_mat_eb object. First column is IID and second column is FID.}

\item{Tune_Y}{The observed phenotypes for the Tune Data.}

\item{Validation_PRS}{PRS's for the Validation Data built using the prs_mat_eb object. First column is IID and second column is FID.}

\item{family}{Distribution of the response, either "Gaussian" or "Binary".}
}
\value{
Named list of Cleaned Tune PRS and Cleaned Validation PRS. First two columns of each data.frame are IID and FID.
}
\description{
Clean the PRS's for the Tune and Validation set for the Super Learner step.
}
