% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_tune.R
\name{PRS_tune}
\alias{PRS_tune}
\title{Select best set of snps for PRS}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, score_farm)
}
\arguments{
\item{x}{Plink binary execute file. Default "plink2".}

\item{r2_vec}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{pthres}{File name and location of written q_range dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'q_range_file'}

\item{wc_base_vec}{scores dataframe produced by PreparePlinkfile().}

\item{pheno_df}{p_values dataframe produced by PreparePlinkfile()}

\item{tune_number}{Number of samples to select from x for tuning}
}
\description{
Select best set of snps for PRS
}
\keyword{PRS,}
\keyword{tuning}
