% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSscoreEBayes.R
\name{PRSscoreEBayes}
\alias{PRSscoreEBayes}
\title{PRSscoreEbayes}
\usage{
PRSscoreEBayes(
  plink2_exec = "plink2 ",
  bfile,
  ebayes = TRUE,
  eb_plink_list,
  plink_list,
  pthres = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.005, 0.05, 0.5, 1),
  threads = 4,
  memory = 8000,
  results_dir,
  out_prefix = as.null(),
  params_farm = as.null()
)
}
\arguments{
\item{plink2_exec}{description}

\item{bfile}{description}

\item{eb_plink_list}{description}

\item{plink_list}{description}

\item{pthres}{Default c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)}

\item{threads}{Default 4}

\item{memory}{Default 8000}

\item{results_dir}{description}

\item{out_prefix}{description}

\item{params_farm}{description}
}
\description{
Performs plink2a --score with EBayes estimates
}
\keyword{clump}
\keyword{plink1.9}
