% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSTrainValSplit.R
\name{PRSTrainValSplit}
\alias{PRSTrainValSplit}
\title{Get the SNP set for estimating the covariance matrix of prior distribution in empirical Bayes method}
\usage{
PRSTrainValSplit(x, n = 0.5)
}
\arguments{
\item{x}{description}

\item{n}{description}
}
\value{
list
}
\description{
Get the SNP set for estimating the covariance matrix of prior distribution in empirical Bayes method
}
