% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunClump.R
\name{RunClump}
\alias{RunClump}
\title{Perform clumping with plink1.9}
\usage{
dimCT(plink19_exec, params_farm)
}
\arguments{
\item{plink19_exec}{Plink binary execute file. Default "plink".}
}
\description{
Perform clumping with plink1.9
}
\keyword{clump}
\keyword{plink1.9}
