% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetClumpFarm.R
\name{SetClumpFarm}
\alias{SetClumpFarm}
\title{set Plink clumping parameters}
\usage{
SetClumpFarm(
  plink19exec = "plink",
  ref_plink,
  target_plink,
  ref_split,
  target_split,
  wc_base_vec = c(50, 100),
  r2_vec = c(0.01, 0.05, 0.1, 0.2, 0.5, 0.8),
  threads = 2,
  mem = 8000
)
}
\arguments{
\item{plink19exec}{path and name to plink1.9 executable. Default "plink"}

\item{ref_plink}{Location and prefix for reference plink binary data set}

\item{target_plink}{Location and prefix for target plink binary data set}

\item{ref_split}{Reference snps data table generated from SplitSum().
Defaults to sum_other_ref.}

\item{target_split}{Target snps data table generated from SplitSum().
Defaults to sum_tar_ref}

\item{wc_base_vec}{Base window size. Defaults to c(50,100)}

\item{r2_vec}{r square vector. Defaults to c(0.01,0.05,0.1,0.2,0.5,0.8)}

\item{threads}{--threads for plink1.9 Defaults to 2}

\item{mem}{--memory for plink1.9 Defaults to 8000 mb}

\item{ref_out}{folder and ref file prefix for plink output}

\item{target_out}{folder and target file prefix for plink output}
}
\value{
A named list
}
\description{
This function sets the plink1.9 clumping parameters
flags values, r square and base window size vectors
}
\examples{
temp.dir <- test/temp.dir
system("mkdir -p test/temp.dir")
clump_farm <- SetClumpFarm(plink19exec = /Apps/plink1.9/plink,
                           ref_plink = "temp/plink/Eur",
                           target_plink = "temp/plink/Afr",
                           ref_split = "temp/Eur_split",
                           target_split = "temp/Afr_split",
                           wc_base_vec = c(50,100),
                           r2_vec = c(0.01,0.05,0.1,0.2,0.5,0.8),
                           threads = 2,
                           mem = 8000,
                           ref_out = 'temp/ref'
                           target_out = 'temp/target'
)
}
\keyword{clump}
\keyword{plink1.9}
