% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitSum.R
\name{SplitSum}
\alias{SplitSum}
\title{Generate the combinations of all the tumor characteristics.}
\usage{
SplitSum(sum_com)
}
\arguments{
\item{sum_com}{The GWAS summary statistics for the target population. The data needs to have following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect allele. BETA is the regression coefficients for linear regression, log-odds ratio for logistic regression. SE is the standard error for BETA.}
}
\value{
GWAS summary statistics for the target population with aligned effects, standard error and p-value for the other population
}
\description{
Generate the combinations of all the tumor characteristics.
}
