% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestModel.R
\name{TestModel}
\alias{TestModel}
\title{Get the SNP set for estimating the covariance matrix of prior distribution in empirical Bayes method}
\usage{
TestModel(x, y, validate, ml_library = c("SL.glmnet", "SL.ridge", "SL.nnet"))
}
\arguments{
\item{x}{description}

\item{y}{validate}

\item{ml_library}{description}
}
\value{
lm model
}
\description{
Get the SNP set for estimating the covariance matrix of prior distribution in empirical Bayes method
}
