% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSplitTables.R
\name{WriteSplitTables}
\alias{WriteSplitTables}
\title{Creates the two data tables created by SplitSum() and writes them to
files used by Plink 1.9 for LD clumping.}
\usage{
WriteSplitTables(x, refFile = "sum", targetFile = "sum")
}
\arguments{
\item{x}{List containing data.tables created by SplitSum()}

\item{refFile}{export path and prefix file name for ref data.table.
Output file will be named "<refFile>_ref"}

\item{targetFile}{export path and prefix file name for target data.table.
Output file will be named "<targetFile>_target"}
}
\description{
Creates the two data tables created by SplitSum() and writes them to
files used by Plink 1.9 for LD clumping.
}
