% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignSum.R
\name{AlignSum}
\alias{AlignSum}
\title{Align GWAS summary stat for the reference population to the target population.}
\usage{
AlignSum(sum_target, sum_ref)
}
\arguments{
\item{sum_target}{The GWAS summary statistics for the target population.}

\item{sum_ref}{The GWAS summary statistics for the reference population.
The most commonly used reference population is European.}
}
\value{
data.frame object with ref population GWAS summary statistics
aligned with target GWAS summary statistics.
}
\description{
Align GWAS summary statistics for the reference population to the target population.
The summary statistics data must have similar formats and include the
following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect
allele. BETA is the regression coefficients for linear regression, log-odds
ratio for logistic regression. SE is the standard error for BETA.
}
\examples{
data <- "data/"
sum_EUR <- fread(paste0(data,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data,"AFR_sumdata.txt"),header=T)
ref_split_file <- "sum_EUR.txt"
target_split_file <- "sum_AFR.txt"

AlignSum(sum_target = sum_AFR,
         sum_ref = sum_EUR)
}
