% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignSum.R
\name{AlignSum}
\alias{AlignSum}
\title{Align GWAS summary stat for the reference population to the target population.}
\usage{
AlignSum(
  sum_target,
  sum_ref,
  results_dir,
  ref_split_file = "temp/sum_ref.txt",
  target_split_file = "temp/sum_target.txt",
  SplitSum = TRUE
)
}
\arguments{
\item{sum_ref}{The GWAS summary statistics for the reference population.
The most commonly used reference population is European.}

\item{ref_split_file}{If SplitSum = TRUE (Default) provide a file path and prefix name
for the reference snps produced from the SplitSum() function.  The default
name "sum" is passed to WriteSplitTables() which creates the file
"sum_ref.txt" in the current working directory. This file is the input for the
plink --clump flag}

\item{target_split_file}{If SplitSum = TRUE (Default) provide a file path and prefix name
for the target snps produced from the SplitSum() function. The default
name "sum" is passed to WriteSplitTables() which creates the file
"sum_target.txt" in the current working directory. This file is the input for
the plink --clump flag}

\item{SplitSum}{Execute the SplitSum() and WriteSplitTables() function.
Default = TRUE.}

\item{sum_tar}{The GWAS summary statistics for the target population.}
}
\value{
data.frame object with ref population GWAS summary statistics
aligned with target GWAS summary statistics. The resulting global variable is
named "sum_com". If SplitSum = TRUE, the SplitSum() and WriteSplitTables()
functions will be executed and the following global variables will be created;
"sum_ref" for the reference snps and "sum_target" for the target snps created
by SplitSum(). The data.frames are also written to files named "sum_ref.txt"
and "sum_target.txt" respectively
}
\description{
Align GWAS summary statistics for the reference population to the target population.
The summary statistics data must have similar formats and include the
following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect
allele. BETA is the regression coefficients for linear regression, log-odds
ratio for logistic regression. SE is the standard error for BETA.
}
\examples{
data.dir <- "data/"
temp.dir = "test/temp.dir/"
sum_EUR <- fread(paste0(data.dir,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data.dir,"AFR_sumdata.txt"),header=T)
ref_split_file <- "sum_EUR.txt"
target_split_file <- "sum_AFR.txt"

AlignSum(sum_target = sum_AFR,
         sum_ref = sum_EUR,
         results_dir = results.dir,
         ref_split_file = ref_split_file,
         target_split_file = target_split_file,
         SplitSum = TRUE)
}
