% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignSumMulti.R
\name{AlignSumMulti}
\alias{AlignSumMulti}
\title{Generate the combinations of all the tumor characteristics.}
\usage{
AlignSumMulti(sum_target, sum_ref_list, ref_names)
}
\arguments{
\item{sum_target}{The GWAS summary statistics for the target population. The data needs to have following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect allele. BETA is the regression coefficients for linear regression, log-odds ratio for logistic regression. SE is the standard error for BETA.}

\item{sum_ref_list}{A list that contains the GWAS summary statistics for multiple populations.}

\item{ref_names}{The names of the other anescestries for GWAS summary statistics}
}
\value{
GWAS summary statistics for the target population with aligned effects, standard error and p-value for the other populations
}
\description{
Generate the combinations of all the tumor characteristics.
}
