% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEBEffectSize.R
\name{CalculateEBEffectSize}
\alias{CalculateEBEffectSize}
\title{CalculateEBEffectSize}
\usage{
CalculateEBEffectSize(
  bfile,
  prs_tune,
  plink_list,
  memory = 8000,
  threads = 2,
  out_prefix,
  results_dir,
  params_farm = as.null()
)
}
\arguments{
\item{bfile}{description}

\item{prs_tune}{description}

\item{plink_list}{description}

\item{memory}{description}

\item{threads}{description}

\item{out_prefix}{description}

\item{results_dir}{description}

\item{params_farm}{list of parameters generated by SetClumpFarm().}

\item{plink2_exec}{description}
}
\description{
Performs the entire Empirical-Bayes estimation of effect sizes
workflow as outlined in step2 of the vignette
}
\keyword{clump}
\keyword{plink1.9}
