% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateQRange.R
\name{CreateQRange}
\alias{CreateQRange}
\title{Create q_range object for PreparePlinkFile()}
\usage{
CreateQRange(x)
}
\arguments{
\item{x}{pthres}
}
\value{
CreateQRange() object
}
\description{
Calls CreateQRange() with either the default pthres or user input values
Not exported
}
