% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimCT.R
\name{dimCT}
\alias{dimCT}
\title{dimCT}
\usage{
dimCT(
  plink19_exec = "plink",
  plink2_exec = "plink",
  results_dir = "./",
  sum_target,
  sum_ref,
  ref_plink,
  target_plink,
  test_target_plink,
  out_prefix = as.null(),
  wc_base_vec = c(50, 100),
  r2_vec = c(0.01, 0.05, 0.1, 0.2, 0.5, 0.8),
  pthres = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.005, 0.05, 0.5, 1),
  threads = 2,
  memory = 8000,
  params_farm = as.null()
)
}
\arguments{
\item{plink19_exec}{Plink binary execute file. Default "plink".}

\item{plink2_exec}{Plink2 binary execute file. Default "plink2".}

\item{results_dir}{Folder or directory to export files}

\item{sum_target}{The GWAS summary statistics for the target population.}

\item{sum_ref}{The GWAS summary statistics for the reference population.}

\item{ref_plink}{Plink binary files for the reference population.}

\item{target_plink}{Plink binary files for the target population.}

\item{test_target_plink}{Plink binary files for the target test population.}

\item{out_prefix}{Prefix for exported files. Recommended when plink files
are divided by chromosome or chunks, i.e., "chr1" or "chr2", and CTSLEB
script is dispatched in parallel.}

\item{wc_base_vec}{Base clumping window size. Defaults to c(50,100)}

\item{r2_vec}{r square vector. Used by Plink1.9 "--clump r^2" parameter.
Defaults to c(0.01,0.05,0.1,0.2,0.5,0.8).}

\item{pthres}{p-value thresholds for plink to create multiple PRSs. Used in Plink2 by "--q-score-range" parameter.
Defaults to c(5e-08,5e-07, 5e-06, 5e-05, 5e-04, 0.005, 0.05, 0.5, 1).}

\item{threads}{Plink1.9 "--threads" parameter. Defaults to 2}

\item{memory}{Plink1.9 "--memory" parameter. Defaults to 8000 mb}

\item{params_farm}{List of plink parameters produced from SetParamsFarm()}
}
\description{
This function performs the entire two-dimensional clumping and thresholding
workflow (dimCT) as outlined in step1 of the vignette
}
\examples{
data <- "data/"
results <- "test/"
temp <- paste0(results,"temp/")
plink19_exec <- "~/Apps/plink_v1.9/plink"
plink2_exec <- "~/Apps/plink2a/plink2"

sum_EUR <- fread(paste0(data,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data,"AFR_sumdata.txt"),header=T)

Eur_ref_plinkfile <- paste0(data,"EUR_ref_chr22")
Afr_ref_plinkfile <- paste0(data,"AFR_ref_chr22")
Afr_test_plinkfile <- paste0(data,"AFR_test_chr22")
outprefix <- "chr22"

PRS_farm <- SetParamsFarm(plink19_exec = plink19_exec,
                          plink2_exec = plink2_exec)
prs_mat <- dimCT(results_dir = results,
                 sum_target = sum_AFR,
                 sum_ref = sum_EUR,
                 ref_plink = Eur_ref_plinkfile,
                 target_plink = Afr_ref_plinkfile,
                 test_target_plink = Afr_test_plinkfile,
                 out_prefix = outprefix,
                 params_farm = PRS_farm)
}
\keyword{clump}
\keyword{plink1.9}
