% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimCT.R
\name{dimCT}
\alias{dimCT}
\title{dimCT}
\usage{
dimCT(plink19_exec, clump_farm)
}
\arguments{
\item{plink19_exec}{Plink binary execute file. Default "plink".}

\item{clump_farm}{list of parameters generated by SetClumpFarm().}
}
\description{
This function performs the entire two-dimensional clumping and thresholding
workflow (dimCT) as outlined in step1 of the vignette
}
\examples{
data.dir <- "data/"
temp.dir <- test/temp.dir
system("mkdir -p test/temp.dir")
sum_EUR <- fread(paste0(data.dir,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data.dir,"AFR_sumdata.txt"),header=T)
Eur_plinkfile <- paste0(data.dir,"EUR_ref_chr22")
Afr_plinkfile <- paste0(data.dir,"AFR_ref_chr22")
ref_out <- "test/temp/ref",
target_out <- "test/temp/target"
clump_farm <- SetClumpFarm(plink19exec = /Apps/plink1.9/plink,
                           wc_base_vec = c(50,100),
                           r2_vec = c(0.01,0.05,0.1,0.2,0.5,0.8),
                           threads = 2,
                           mem = 8000,
)
dimCT(plink19_exec=plink19_exec,
      sum_target = sum_AFR,
      sum_ref = sum_EUR,
      ref_plink=ref_plink,
      target_plink=target_plink,
      ref_clump_out=ref_clump_out,
      target_clump_out=target_clump_out,
      PRS_farm)
}
\keyword{clump}
\keyword{plink1.9}
