% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBayesPostMean.R
\name{EBayesPostMean}
\alias{EBayesPostMean}
\title{Estimate the EB posterior mean}
\usage{
EBayesPostMean(clump_info, snp_set)

EBayesPostMean(x,y)
}
\arguments{
\item{x}{unique_infor data.frame from PreparePlinkFile()}

\item{y}{SNP data.frame from GetSNPSet()}
}
\description{
Estimate the EB posterior mean for SNPs in the unique_infor data.frame from
PreparePlinkFile()
}
