% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBpostMulti.R
\name{EBpostMulti}
\alias{EBpostMulti}
\title{Estimate Bayesian posterior meanusing the empirical Bayes algorithm}
\usage{
EBpostMulti(unique_infor, SNP_set, sum_com, other_ans_names)
}
\arguments{
\item{unique_infor}{the unique_infor from PreparePlinkFile output, which contains information for all SNPs after clumping step}

\item{SNP_set}{the SNP set from CT step for estimating the covariance matrix for the prior distribution}

\item{sum_com}{summary statistics for all ancestries after allele alignment}

\item{other_ans_names}{the other ancestries' name used in the analyses}
}
\value{
Bayesian postier mean for all SNPs after-clumping
}
\description{
Estimate Bayesian posterior meanusing the empirical Bayes algorithm
}
