% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatePriorMulti.R
\name{EstimatePriorMulti}
\alias{EstimatePriorMulti}
\title{Estimate the covariance matrix for the prior distribution used in EB step}
\usage{
EstimatePriorMulti(snp_set, ref_names, sum_com)
}
\arguments{
\item{snp_set}{the SNP set from CT step for estimating the covariance matrix for the prior distribution}

\item{ref_names}{description}

\item{sum_com}{AlignSumMulti() object}
}
\value{
the covariance matrix estimate
}
\description{
Estimate the covariance matrix for the prior distribution used in EB step
}
