% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSNPSet.R
\name{GetSNPSet}
\alias{GetSNPSet}
\title{Get the SNP set for estimating the covariance matrix of prior distribution in empirical Bayes method}
\usage{
GetSNPSet(snp_ind, scores, clump_info)
}
\arguments{
\item{snp_set_ind}{the SNP set index}

\item{score_file}{the score_file from PreparePlinkFile output}

\item{unique_infor}{the unique_infor from PreparePlinkFile output}
}
\value{
the SNP set
}
\description{
Get the SNP set for estimating the covariance matrix of prior distribution in empirical Bayes method
}
