% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSNPSet.R
\name{GetSNPSet}
\alias{GetSNPSet}
\title{Get the best performance SNPS}
\usage{
GetSNPSe(snp_id, scores, clump_info)
}
\arguments{
\item{scores}{The scores dataframe from PreparePlinkFile output.}

\item{clump_info}{The unique_infor from PreparePlinkFile output.}

\item{snp_set_ind}{Column name from prs_mat (output from PRSscore()) with best
performance against the tuning data set.}
}
\value{
data.frame of SNP performance metrics
}
\description{
Get the SNP set for estimating the covariance matrix of prior distribution in
empirical Bayes method
}
