% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_combine_PRS.R
\name{helper_combine_PRS}
\alias{helper_combine_PRS}
\title{Helper function for PRSscore()}
\usage{
helper_combine_PRS(scores, pthres, params_farm = as.null())
}
\value{
'prs_mat' global variable. Matrix of combined PRSs with clumping and
pthres as column name
}
\description{
Combines PRS
Not exported
}
