% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_return_list.R
\name{helper_return_list}
\alias{helper_return_list}
\title{Helper function for PreparePlinkFile()}
\usage{
helper_return_list(x, output, score_file, p_value_file, unique_infor, q_range)
}
\arguments{
\item{x}{return_list}

\item{output}{path to output folder}
}
\value{
CreateQRange() object
}
\description{
Return either a list containing the score_file, p_value_file, unique_infor and
q_range or the global variables for the same objects and write the files.
Not exported
}
