% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_SplitSum.R
\name{helper_SplitSum}
\alias{helper_SplitSum}
\title{Helper function for AlignSum()}
\usage{
helper_SplitSum(x, sum_com, ref_split_file, target_split_file)
}
\arguments{
\item{x}{TRUE default}

\item{ref_split_file}{If SplitSum = TRUE, this is the file path and name for
the reference snps produced from the SplitSum() function.  The default name
"sum_ref.txt" is passed to WriteSplitTables().}

\item{target_split_file}{If SplitSum = TRUE, this is the file path and name
for the target snps produced from the SplitSum() function. The default
name "sum_target.txt" is passed to WriteSplitTables().}
}
\value{
SplitSum() and WriteSplitTables() objects and files
}
\description{
SplitSum = TRUE, call the SplitSum() and WriteSplitTables() functions and
create the sum_com global variable object.
SplitSum = FALSE, only create the sum_com global variable object.
Not exported
}
