% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2score.R
\name{plink2score}
\alias{plink2score}
\title{Generate PRS scores with Plink2}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, score_farm)
}
\arguments{
\item{plink2_exec}{Plink binary execute file. Default "plink2".}

\item{bfile}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{p_value_file}{File name and location of written p_values_temp dataframe
either produced by helper_score_loop() or manually as indicated in the
vignette section 'Generate plink2 PRS'}

\item{q_range_file}{File name and location of written q_range dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'q_range_file'}

\item{score_col_nums}{number of columns in scores dataframe produced by
PreparePlinkfile().}

\item{scores_file}{File name and location of written scores dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'scores_file'}

\item{threads}{maximum number of concurrent threads}

\item{memory}{primary workspace memory}

\item{out}{output folder and file prefix for Plink2. Default name prs_p_other_
<pthres column number>.<range name>.sscore}

\item{p_values}{p_values dataframe produced by PreparePlinkfile()}
}
\value{
plink2 score files
}
\description{
Performs PRS scoring with plink2 for all combinations of clumpings and p-value
threshold combinations
}
\keyword{plink2}
\keyword{score}
