% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlinkFile.R
\name{PreparePlinkFile}
\alias{PreparePlinkFile}
\title{Prepare the files for PLINK2 to calculate PRSs}
\usage{
PreparePlinkFile(
  snp_list = snp_list,
  sum_com,
  pthres = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.005, 0.05, 0.5, 1),
  results_dir,
  params_farm = as.null()
)
}
\arguments{
\item{snp_list}{the snp_list result from the two-dimensional clumping}

\item{sum_com}{the sum_com result from the AlignSum function.}

\item{pthres}{vector of p-value thresholds. Default
c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)}

\item{results_dir}{description}
}
\value{
Creates either the global variables 'scores', 'p_values',
'q_range' and 'unique_infor' or the variable 'plink_list' which contains the
four dataframes in a list
}
\description{
Calls helper_return_list() and helper_CreateQRange()
}
