% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlinkFile.R
\name{PreparePlinkFile}
\alias{PreparePlinkFile}
\title{Prepare the files for PLINK2 to calculate PRSs}
\usage{
PreparePlinkFile(snp_list, sum_com)
}
\arguments{
\item{snp_list}{the snp_list result from the two-dimensional clumping}

\item{sum_com}{the sum_com result from the AlignSum function.}
}
\value{
the score_file, p_value_file and p_other_file for the two-dimensional thresholding
}
\description{
Prepare the files for PLINK2 to calculate PRSs
}
