% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlinkFile.R
\name{PreparePlinkFile}
\alias{PreparePlinkFile}
\title{Prepare the files for PLINK2 to calculate PRSs}
\usage{
PreparePlinkFile(
  snp_list = snp_list,
  sum_com,
  pthres = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.005, 0.05, 0.5, 1),
  results_dir,
  return_list = FALSE,
  params_farm = as.null()
)
}
\arguments{
\item{snp_list}{the snp_list result from the two-dimensional clumping}

\item{sum_com}{the sum_com result from the AlignSum function.}

\item{pthres}{vector of p-value thresholds. Default
c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)}

\item{return_list}{TRUE will return a 'plink_files' list containing the
plink_files[1]"scores" plink_files[2]"p_values", plink_files[3]"unique_infor"
and plink_files[4]"q_range" objects. FALSE will create the global variables for
the same objects and write the q-range and scores objects to tables named
"q_range_file" and "scores_file" respectively. The location of these tables
will be stored as global variables named  "q_range_file" and "scores_file"
respectively. Default return_list=FALSE}
}
\value{
Creates either the global variables 'scores', 'p_values',
'q_range' and 'unique_infor' or the variable 'plink_list' which contains the
four dataframes in a list
}
\description{
Calls CreateQRange() using either the default pthres or user input values
}
