% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePlinkFileEBayes.R
\name{PreparePlinkFileEBayes}
\alias{PreparePlinkFileEBayes}
\title{PreparePlinkFileEBayes}
\usage{
PreparePlinkFileEBayes(
  snp_list,
  clump_info,
  post_clump_info,
  post_beta,
  results_dir
)
}
\arguments{
\item{snp_list}{description}

\item{clump_info}{description}

\item{post_clump_info}{description}

\item{post_beta}{description}

\item{results_dir}{description}
}
\description{
This function performs the entire two-dimensional clumping and thresholding
workflow (dimCT) as outlined in step1 of the vignette
}
\keyword{clump}
\keyword{plink1.9}
