% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSscore.R
\name{PRSscore}
\alias{PRSscore}
\title{PRSscore}
\usage{
PRSscore(plink2_exec, bfile, plink_list, pthres, memory, out, params_farm)
}
\arguments{
\item{plink2_exec}{Plink binary execute file. Default "plink2".}

\item{bfile}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{pthres}{vector of p_value thresholds used to select snps for PRS.
Default vector name is 'pthres' which is generated by PreparePlinkfile()}

\item{threads}{plink parameter for maximum number of concurrent threads}

\item{memory}{plink parameter for primary workspace memory}

\item{results_dir}{Folder or directory to export files}

\item{out_prefix}{Prefix for exported files. Recommended when plink files
are divided by chromosome or chunks, i.e., "chr1" or "chr2", and CTSLEB
script is dispatched in parallel}

\item{params_farm}{params_farm object name}

\item{file_list}{description}
}
\description{
Master function for scoring with plink2.
If additional options are required, please use the source() command
as demonstrated in the vignette.
}
