% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSscore.R
\name{PRSscore}
\alias{PRSscore}
\title{PRSscore}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, params_farm)
}
\arguments{
\item{plink2_exec}{Plink binary execute file. Default "plink2".}

\item{bfile}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{plink_list}{description}

\item{pthres}{vector of p_value thresholds used to select snps for PRS.
Default vector name is 'pthres' which is generated by PreparePlinkfile()}

\item{threads}{maximum number of concurrent threads}

\item{memory}{primary workspace memory}

\item{results_dir}{description}

\item{out_prefix}{description}

\item{params_farm}{params_farm object name}
}
\description{
Master function for scoring with plink2.
If additional options are required, please use the source() command
as demonstrated in the vignette.
}
\examples{
pthres <- c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)

PRSscore(plink2_exec = "plink2 ",
            bfile = bfile,
            plink_list = plink_list,
            pthres = pthres,
            threads = 4,
            memory = 8000,
            out = prs_out,
            params_farm = params_farm)
}
\keyword{plink2}
\keyword{score}
