% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSscore.R
\name{PRSscore}
\alias{PRSscore}
\title{PRSscore}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, score_farm)
}
\arguments{
\item{plink2_exec}{Plink binary execute file. Default "plink2".}

\item{bfile}{prefix for target plink binary set (prefix.bed, prefix.bim, prefix.fam)}

\item{q_range_file}{File name and location of written q_range dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'q_range_file'}

\item{scores}{scores dataframe produced by PreparePlinkfile().}

\item{scores_file}{File name and location of written scores dataframe produced
by PreparePlinkfile(). By default PreparePlinkfile() names this 'scores_file'}

\item{p_values}{p_values dataframe produced by PreparePlinkfile()}

\item{pthres}{vector of p_value thresholds used to select snps for PRS.
Default vector name is 'pthres' which is generated by PreparePlinkfile()}

\item{threads}{maximum number of concurrent threads}

\item{memory}{primary workspace memory}

\item{out}{output folder and file prefix for Plink2. Default name prs_p_other_
<pthres column number>.<range name>.sscore}
}
\description{
Master function for scoring with plink2.
If additional options are required, please use the source() command
as demonstrated in the vignette.
}
\examples{
pthres <- c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)

PreparePlinkFile(snp_list = snp_list,
                 sum_com = sum_com,
                 pthres = pthres,
                 return_list = FALSE)

prs_out <- paste0(temp.dir,"prs_p_other")
PRSscore(plink2_exec = "plink2 ",
            bfile = bfile,
            q_range_file = q_range_file,
            scores = scores,
            scores_file = scores_file,
            p_values = p_values,
            pthres = pthres,
            threads = 4,
            memory = 8000,
            out = prs_out,
            scoreFarm = scoreFarm)
}
\keyword{plink2}
\keyword{score}
