% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSscoreEBayes.R
\name{PRSscoreEBayes}
\alias{PRSscoreEBayes}
\title{Plink --score wrapper}
\usage{
PRSscoreEBayes <- function(plink2_exec,bfile,ebayes, eb_plink_list,
plink_list, pthres, threads, memory, results_dir, out_prefix, params_farm)
}
\arguments{
\item{plink2_exec}{Plink2 binary execute file. Default "plink2".}

\item{bfile}{Plink binary files for the target test population.}

\item{eb_plink_list}{List of plink data frame and files produced from
PreparePlinkFileEBayes()}

\item{plink_list}{List of plink data.frame and files from PreparePlinkFiles()}

\item{pthres}{P value thresholds. Defaults to
c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0)}

\item{threads}{Plink "--threads" parameter. Defaults to 2}

\item{memory}{Plink "--memory" parameter. Defaults to 8000 mb}

\item{results_dir}{Folder or directory to export files.}

\item{out_prefix}{Prefix for exported files. Recommended when plink files
are divided by chromosome or chunks, i.e., "chr1" or "chr2", and CTSLEB
script is dispatched in parallel.}

\item{params_farm}{List of plink parameters produced from SetParamsFarm()}
}
\description{
Wrapper to run plink2a --score routine with EBayes estimates and then combine
all the PRSs.
}
\keyword{plink2}
