% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetClumpFarm.R
\name{SetClumpFarm}
\alias{SetClumpFarm}
\title{set Plink clumping parameters}
\usage{
SetClumpFarm(
  plink19exec = "plink",
  sum_ref,
  sum_target,
  ref_plink,
  target_plink,
  refFile = "sum_ref.txt",
  targetFile = "sum_target.txt",
  wc_base_vec = c(50, 100),
  r2_vec = c(0.01, 0.05, 0.1, 0.2, 0.5, 0.8),
  threads = 2,
  mem = 8000,
  ref_out = "temp/ref",
  target_out = "temp/target"
)
}
\arguments{
\item{plink19exec}{path and name to plink1.9 executable. Default "plink"}

\item{sum_ref}{The GWAS summary statistics for the reference population.
The most commonly used reference population is European.}

\item{ref_plink}{Location and prefix for reference plink binary data set}

\item{target_plink}{Location and prefix for target plink binary data set}

\item{refFile}{Path and prefix file name for reference snps generated by
SplitSum(). This is the input file for the Plink1.9 --clump flag. Defaults to
sum_ref.txt}

\item{targetFile}{Path and prefix file name for target snps generated by
SplitSum(). This is the input file for the Plink1.9 --clump flag. Defaults to
sum_target.txt}

\item{wc_base_vec}{Base window size. Defaults to c(50,100)}

\item{r2_vec}{r square vector. Defaults to c(0.01,0.05,0.1,0.2,0.5,0.8)}

\item{threads}{--threads for plink1.9 Defaults to 2}

\item{mem}{--memory for plink1.9 Defaults to 8000 mb}

\item{ref_out}{folder and ref file prefix for plink output}

\item{target_out}{folder and target file prefix for plink output}

\item{sum_tar}{The GWAS summary statistics for the target population.}
}
\value{
A named list
}
\description{
This function sets the plink1.9 clumping parameters
flags values, r square and base window size vectors
}
\examples{
data.dir <- "data/"
temp.dir <- test/temp.dir
system("mkdir -p test/temp.dir")
sum_EUR <- fread(paste0(data.dir,"EUR_sumdata.txt"),header=T)
sum_AFR <- fread(paste0(data.dir,"AFR_sumdata.txt"),header=T)
Eur_plinkfile <- paste0(data.dir,"EUR_ref_chr22")
Afr_plinkfile <- paste0(data.dir,"AFR_ref_chr22")
clump_farm <- SetClumpFarm(plink19exec = /Apps/plink1.9/plink,
                           sum_ref = sum_EUR,
                           sum_target = sum_AFR,
                           ref_plink = Eur_plinkfile,
                           target_plink = Afr_plinkfile,
                           refFile = "test/temp/Eur",
                           targetFile = "test/temp/Afr",
                           wc_base_vec = c(50,100),
                           r2_vec = c(0.01,0.05,0.1,0.2,0.5,0.8),
                           threads = 2,
                           mem = 8000,
                           ref_out = "test/temp/ref",
                           target_out = "test/temp/target"
)
}
\keyword{clump}
\keyword{plink1.9}
