% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetParamsFarm.R
\name{SetParamsFarm}
\alias{SetParamsFarm}
\title{set Plink clumping parameters}
\usage{
SetParamsFarm(
  plink19_exec = "plink",
  plink2_exec = "plink2",
  wc_base_vec = c(50, 100),
  r2_vec = c(0.01, 0.05, 0.1, 0.2, 0.5, 0.8),
  pthres = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.005, 0.05, 0.5, 1),
  threads = 2,
  mem = 8000
)
}
\arguments{
\item{plink19_exec}{path and name to plink1.9 executable. Default "plink"}

\item{plink2_exec}{path and name to plink2 executable. Default "plink2"}

\item{wc_base_vec}{Base window size. SetClumpFarm will automaticaly create the
global 'wc_base_vec' since it is required in other workflow sections. Defaults
to c(50,100)}

\item{r2_vec}{r square vector. SetClumpFarm will automaticaly create the
global 'r2_vec' since it is required in other workflow sections. Defaults to
c(0.01,0.05,0.1,0.2,0.5,0.8).}

\item{pthres}{description}

\item{threads}{--threads for plink1.9 Defaults to 2}

\item{mem}{--memory for plink1.9 Defaults to 8000 mb}

\item{ref_out}{folder and ref file prefix for plink output}

\item{target_out}{folder and target file prefix for plink output}
}
\value{
A named list
}
\description{
This function sets the plink1.9 clumping and plink2a thresholding parameters
and flags
}
\examples{
params_farm <- SetParamsFarm(plink19exec = /Apps/plink1.9/plink,
                           wc_base_vec = c(50,100),
                           r2_vec = c(0.01,0.05,0.1,0.2,0.5,0.8),
                           pthres = c(5E-08,5E-07,5E-06,5E-05,5E-04,5E-03,5E-02,5E-01,1.0),
                           threads = 2,
                           mem = 8000
)
}
\keyword{clump}
\keyword{plink1.9}
