% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitSum.R
\name{SplitSum}
\alias{SplitSum}
\title{Inputs an AlignSum() object and splits SNPs into two groups}
\usage{
SplitSum(x)
}
\arguments{
\item{x}{sum_com data.frame object produced by AlignSum().}
}
\value{
A list with the two data.frame objects of SNPs with p-values in the
ref population smaller than the target population in the [[1]] slot.
SNPs with p-values in the target population smaller than the ref population
in the [[2]] slot.
}
\description{
Split the SNPs into two groups: 1. SNPs with p-values in the ref population
smaller than the target population. 2. SNPs with p-values in the target
population smaller than the ref population.
}
\examples{
sum_com <- AlignSum(sum_tar = sum_AFR, sum_other = sum_EUR)
split_list <- SplitSum(sum_com)
}
