% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitSum.R
\name{SplitSum}
\alias{SplitSum}
\title{Generate the combinations of all the tumor characteristics.}
\usage{
SplitSum(sum_com)
}
\arguments{
\item{sum_com}{The GWAS summary statistics for the target population. The data needs to have following columns at least: CHR, SNP, BP, A1, BETA, SE, P. A1 is the effect allele. BETA is the regression coefficients for linear regression, log-odds ratio for logistic regression. SE is the standard error for BETA.}
}
\value{
GWAS summary statistics for the target population with aligned effects, standard error and p-value for the other population
}
\description{
Generate the combinations of all the tumor characteristics.
}
\examples{

#this is a simulated breast cancer example
#there are around 5000 breast cancer cases and 5000 controls, i.e. people without disease
data[1:5,]
#four different tumor characteristics were included, ER (positive vs negative), PR (positive vs negative), HER2 (positive vs negative), grade (ordinal 1, 2, 3)
#the phenotype file
y <- data[,1:5]
#generate the combinations of all the subtypes
#by default, we remove all the subtypes with less than 10 cases
z.standard <- GenerateZstandard(y)
}
