% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Super_split.R
\name{Super_split}
\alias{Super_split}
\title{Create tune and validation data sets}
\usage{
Plink19Clump(plink2_exec, bfile, q_score_range, score_col_nums,
score, threads, memory, out, score_farm)
}
\arguments{
\item{x}{matrix of PRSs based on EB coefficients produced by PRSscoreEBayes()
or CalculateEBEffectSize()}

\item{y}{vector containing tune phenotypes. This vector should be the same
length}

\item{params_farm}{params farm}

\item{n}{fraction of samples which will be used for tuning. The remainder
will be used for validation. Default 0.5}
}
\description{
Create tune and validation data sets
}
\keyword{PRS,}
\keyword{tuning}
