% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSplitTables.R
\name{WriteSplitTables}
\alias{WriteSplitTables}
\title{Creates the two data tables created by SplitSum()}
\usage{
WriteSplitTables(x, refFile = "sum", targetFile = "sum")
}
\arguments{
\item{x}{List containing data.frame objects created by SplitSum()}

\item{refFile}{export path and prefix file name for ref data.frame
Output file will be named "<refFile>_ref"}

\item{targetFile}{export path and prefix file name for target data.frame
Output file will be named "<targetFile>_target"}
}
\value{
The ref and target aligned data.frames produced by SplitSum() are
written to default files named "sum_ref.txt" and "sum_target.txt" respectively
}
\description{
Creates two data.frame objects from SplitSum() result and writes them to
files. These files are inputs to the Plink 1.9 --clump flag for LD clumping.
}
\examples{
sum_com_split <- SplitSum(sum_com)
WriteSplitTables(x=sum_com_split, refFile = "sum", targetFile = "sum")
}
